requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dominate==2.7.0
et-xmlfile==1.1.0
Flask==2.3.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.3.0
WTForms==3.0.1
